update cypress test
