+++ cypress/e2e/plugin-example.cy.ts
@@ -1,9 +1,12 @@
+import { expectElementExists } from './utils/expects';
+import { getSomeComponent } from './utils/getters';
 import { loadPlayerAndSetMedia } from './utils/setup';
 
 describe('plugin example', () => {
   it('should load the plugin', () => {
     loadPlayerAndSetMedia({}).then((player) => {
       expect(player).to.exist;
+      expectElementExists(getSomeComponent);
     });
   });
 });
