+++ cypress/e2e/utils/expects.ts
@@ -1,0 +1,15 @@
+export const expectElementExists = (getElement: () => any) => {
+  getElement().should('exist');
+};
+
+export const expectElementDoesntExist = (getElement: () => any) => {
+  getElement().should('not.exist');
+};
+
+export const expectElementVisible = (getElement: () => any) => {
+  getElement().should('be.visible');
+};
+
+export const expectElementNotVisible = (getElement: () => any) => {
+  getElement().should('not.be.visible');
+};
