+++ package.json
@@ -43,7 +43,7 @@
     "cy:run:firefox": "cypress run --browser firefox --config video=false",
     "cy:run:edge": "cypress run --browser edge --config video=false",
     "cy:run:safari": "cypress run --browser webkit --config video=false",
-    "test:prepare": "yarn run build && copyfiles -f ./dist/playkit-detect-ad-block.js ./dist/playkit-detect-ad-block.js.map ./cypress/public",
+    "test:prepare": "yarn run build && copyfiles -f ./dist/playkit-plugin-example.js ./dist/playkit-plugin-example.js.map ./cypress/public",
     "test:watch": "yarn run test:prepare && yarn run cy:open",
     "test": "yarn run test:prepare && yarn run cy:run"
   },
