+++ src/components/some-component/index.ts
@@ -1,0 +1,1 @@
+export { SomeComponent } from './some-component';
