+++ src/components/some-component/some-component.scss
