+++ src/plugin-example.tsx
@@ -1,6 +1,6 @@
 import { h } from 'preact';
 import { BasePlugin, ui } from '@playkit-js/kaltura-player-js';
-import { SomeWrappedComponent } from './components/more-icon/some-component.component';
+import { SomeComponent } from './components/some-component/';
 import { PluginExampleConfig } from './types';
 
 export const pluginName = 'pluginExample';
@@ -25,7 +25,7 @@ export class PluginExample extends BasePlugin {
       label: 'plugin-example',
       area: ui.ReservedPresetAreas.InteractiveArea,
       presets: [ui.ReservedPresetNames.Playback, ui.ReservedPresetNames.Live],
-      get: () => <SomeWrappedComponent someTitle={this.config.someTitle} />
+      get: () => <SomeComponent someTitle={this.config.someTitle} />
     });
   }
